updated stuff to fix stuff

+++ core/dto.py
@@ -2,6 +2,7 @@ from typing import List, Optional
 from dataclasses import dataclass
 from datetime import datetime
 
+
 @dataclass
 class Character:
     name: str
@@ -9,7 +10,7 @@ class Character:
     region: str
     char_class: str
     spec_name: str
-    item_level: int
+    item_level: float
     profile_url: str
     thumbnail_url: str
     last_crawled_at: datetime
